[Feature] Allow paths to be loaded into entity factory (#343)
Adds a load method to the entity factory so that factories can
be loaded from additional paths. This is particularly useful if
using entity factories exist in a package rather than a Laravel
app.

Additionally this pattern matches the behaviour of the Eloquent
factory, which allows additional paths to be loaded via a load
method.